% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_module.R
\name{fun2mod}
\alias{fun2mod}
\title{Function to module}
\usage{
fun2mod(fun, assign_to_global_env = TRUE)
}
\arguments{
\item{function}{The name of an R function to be converted into a module}
}
\value{
Shiny module components, UI and server, as functions in `.GlobalEnv`. The module is also returned as text and can be captured in a variable.
}
\description{
Convert a function into a shiny module.
}
\details{
The arguments of the function become input widgets. The output of the function is passed onto the appropriate `render*` function: `renderPlot` for plot, `renderTable` for tables and `renderText` for text.
}
\examples{

# First we define a function
sum_ <- function(x, y) x + y

# Second we convert to module
mod <- fun2mod(fun = "sum_")

# Third we launch a shiny app with this module
library(shiny)

ui <- fluidPage(
 sum__ui(id = "A")
)

server <- function(input, output, session) {
 callModule(sum__srv, id = "A")
}

shinyApp(ui, server)

}
